updated the code to use /saveControls for control route
