+++ Servers/routes/control.route.ts
@@ -10,6 +10,7 @@ import {
 } from "../controllers/control.ctrl";
 
 import authenticateJWT from "../middleware/auth.middleware";
+import { saveControls } from "../controllers/control.ctrl";
 
 // GET requests
 router.get("/", /*authenticateJWT,*/ getAllControls);
@@ -17,6 +18,7 @@ router.get("/:id", /*authenticateJWT,*/ getControlById);
 
 // POST, PUT, DELETE requests
 router.post("/", /*authenticateJWT,*/ createControl);
+router.post("/saveControls", /*authenticateJWT,*/ saveControls);
 router.put("/:id", /*authenticateJWT,*/ updateControlById);
 router.delete("/:id", /*authenticateJWT,*/ deleteControlById);
